requirements.txt
flask
flask-socketio
pandas
numpy
statsmodels
scipy
websocket-client
python-dateutil
plotly